Adjust exclusivity for knight styles, auto-clear stale fighting forms
Stale fighting forms (defined as "fighting form you have active but restricted") are cleared at the top of the turn now. This prints a brief message "You readjust your stance".

Knight fighting forms are now in two chunks - the basic 4 are mutually exclusive with each other, and ditto for the xp14 3, but not with the opposite pool. So you can shield bash + eldritch, for example.