Research nerf/tweak: Requires at least 6 int to do experiments.
Replacement for the threat of death by brainlessness that was removed very early on.

This requires that you raise your int score to recover from the damage you do to yourself.

Without such a requirement, the optimum strategy is to chill at 3 int while doing your research, since hits to int that would take you below 3 are "free"