Nymph carry cap idea:
NPC nymphs can carry unlimitted amounts of weight (as long as it's not boulders).

Apply a "nymph magic" weight reduction to each individual stack of items in the PC's inventory, equal to 5x the PC's level.

In the early game this helps with light objects, at the end of the game (when the reduction is 150 aum) you will only actually be carrying your bag and perhaps primary weapon and armor.

Even 5x is maybe too strong.

This separates nymphs from orcs (level-based bonus to final carry cap) and androids and chiros (high natural carry cap all game).